% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advertising.R
\docType{data}
\name{advertising}
\alias{advertising}
\title{Data on different advertising channels and sales}
\format{
A data frame with 200 rows and 5 variables:
\describe{
\item{ID}{id of observation, num}
\item{TV}{Expenditures on TV advertising, num}
\item{radio}{Expenditures on radio advertising, num}
\item{newspaper}{Expenditures on newspaper advertising, num}
\item{sales}{Sales income, num}

}
}
\source{
https://www.statlearning.com/resources-second-edition
}
\usage{
advertising
}
\description{
This is an example dataset from the literature on in statistical learning. It is
analyzed in a book called Introduction to Statistical Learning by James, Witten, Hastie, and Tibshirani.
The dataset consists of 200 samples with four variables: Advertising budgets  for newspaper,
TV, and radio (the predictors) Sales (the outcome)
}
\keyword{datasets}
