% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arthritis.R
\docType{data}
\name{arthritis}
\alias{arthritis}
\title{Data from a clinical trial on an arthritis medication}
\format{
A data frame with 906 rows and 9 variables:
\describe{
\item{id}{Subject Id, num}
\item{y}{self-assessment of rheumatois arthritis five levels 1=poor, ...5= very good, num}
\item{sex}{Sex, fct with two levels make, female}
\item{age}{Age, num}
\item{trt}{Treatment, fact with two levels drug, placebo}
\item{baseline}{Baseline, num with five levels}
\item{time}{Time of self assessment, chr}
\item{baselinescore}{Baselinescore, fct with 5 levels}
\item{score}{Score, fct, 5 levels}
}
}
\source{
atable: Create Tables for Clinical Trial Reports, Armin Ströbel, The R Journal Vol. 11/01, June 2019
https://journal.r-project.org/archive/2019/RJ-2019-001/RJ-2019-001.pdf
}
\usage{
arthritis
}
\description{
The dataset arthritis is from the package multgee Touloumis (2015). This
dataset contains observations of the self-assessment score of arthritis, an ordered variable with five
categories, collected at baseline and three follow-up times during a randomised comparative study of
alternative treatments of 302 patients with rheumatoid arthritis.
The data came from a randomized clinical trial that aimed to evaluate
the effectiveness of the drug Auranofin versus the placebo therapy for the treatment of
rheumatoid arthritis. The five-level (1=poor, . . . , 5=very good) ordinal
multinomial response variable was the self-assessment of rheumatoid arthritis recorded at one
(t = 1), three (t = 2) and five (t = 3) follow-up months.
}
\keyword{datasets}
