% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/college_years_and_starting_salary_data.R
\docType{data}
\name{col_sal}
\alias{col_sal}
\title{Years in college and starting salaries.}
\format{
A data frame with 24 rows and 4 variables:
\describe{
\item{Student.ID}{Identifier for student, chr}
\item{Major}{Subject in which student took his or her major, chr}
\item{Time}{Time, num, number of years of study for given major}
\item{Salary}{Starting salary, num, strating salary in 1000 Euro in the first
job after completion of studies}
}
}
\source{
\url{https://magazine.amstat.org/videos/TYC2020/RoxyPeck.mp4}
}
\usage{
col_sal
}
\description{
Random sample of 24 German students reporting subject major, study years and starting salaries for each
of the 24 anonymous student IDs
}
\keyword{datasets}
