% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dem_dat_by_age_hist.R
\docType{data}
\name{dem_dat_by_age_hist}
\alias{dem_dat_by_age_hist}
\title{Demographic data for countries around the world from the International Database (IBD) of the US
Census by age}
\format{
A data frame with 403152 rows and 8 variables:
\describe{
\item{Country}{Name of country}
\item{ISO2}{ISO 3166 Alpha2 country code}
\item{Year}{Year of observation}
\item{Age}{Age buckets in years}
\item{MR}{Mortality rate, both sexes}
\item{POP}{Population}
\item{DEPND}{Dependency ratio}
\item{BIRTHS}{Births to mothers}
}
}
\source{
US Census International Database
https://www.census.gov/programs-surveys/international-programs/about/idb.html
file idb5yr.txt
}
\usage{
dem_dat_by_age_hist
}
\description{
This dataset contains demographic variables for countries around the world over time starting
from 2024 and reaching up to 2100 broken down by ager groups of 5 year periods.
The dataset has the extension hist to indicate that these
are historical observations. There is a parallel dataset to this one with the extension proj to
indicate that the values are historical observations..
}
\keyword{datasets}
