% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dem_dat_by_sex_and_age_proj.R
\docType{data}
\name{dem_dat_by_sex_and_age_proj}
\alias{dem_dat_by_sex_and_age_proj}
\title{Demographic data for countries around the world from the International Database (IBD) of the US
Census by age}
\format{
A data frame with 767316 rows and 8 variables:
\describe{
\item{Country}{Name of country}
\item{ISO2}{ISO 3166 Alpha2 country code}
\item{Year}{Year of observation}
\item{Sex}{Sex, male, female}
\item{Age}{Age, buckets in years}
\item{MR}{Mortality rate, per 1000}
\item{POP}{Population}
\item{ASFR}{Age specific fertility rate per women}
}
}
\source{
US Census International Database
https://www.census.gov/programs-surveys/international-programs/about/idb.html
file idb5yr.txt
}
\usage{
dem_dat_by_sex_and_age_proj
}
\description{
This dataset contains demographic variables for countries around the world over time starting
from 2024 and reaching up to 2100 broken down by sex and age.
The dataset has the extension proj to indicate that these
are projections. There is a parallel dataset to this one with the extension hist to
indicate that the values are historical observations..
}
\keyword{datasets}
