% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Co2_emissions_vs_gdp_per_capita.R
\docType{data}
\name{emissions}
\alias{emissions}
\title{co2_emissions_versus_gdp_per_capita}
\format{
A data frame with 13769 rows and 5 variables:
\describe{
\item{Entity}{Country name, chr}
\item{Code}{ISO3 code of country, chr}
\item{Year}{Year, num, year of observation}
\item{CO2pc}{Co2 emissions per capita in a year, num}
\item{GDPpc}{GDP per capita in a year, num}
}
}
\source{
\url{https://ourworldindata.org/}
}
\usage{
emissions
}
\description{
Data on co2 emissions per capita versus gdp per capita from 1950 to 1918
for countries around the globe
}
\keyword{datasets}
