% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy_consumption_per_capita.R
\docType{data}
\name{energy_consumption_per_capita}
\alias{energy_consumption_per_capita}
\title{Energy consumption per capita.}
\format{
A data frame with 9833 rows and 4 variables:
\describe{
\item{Country}{Country name, chr}
\item{Code}{ISO3 code of country, chr}
\item{Year}{Year, num, year of observation}
\item{Cons}{Consumption of energy per capita in kwh per person per year, num}
}
}
\source{
\url{https://ourworldindata.org/}
}
\usage{
energy_consumption_per_capita
}
\description{
Data on primary energy consumption per capita measured in kwh per person per year for various countries
around the globe from 1965 - 2021
}
\keyword{datasets}
