% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gehlbach.R
\docType{data}
\name{gehlbach}
\alias{gehlbach}
\title{Data from the evidence of Stephen Gehlbach in United States versus Kristen Gilbert}
\format{
A data frame with 30 rows and 5 variables:
\describe{
\item{year}{year, num}
\item{shift}{shift, fct}
\item{start}{start of shift, date}
\item{end}{end of shift, date}
\item{number_of_deaths}{number of deaths, num}

}
}
\source{
https://books.google.at/books?id=28dQNQAACAAJ
}
\usage{
gehlbach
}
\description{
These are data from hospital records at the Veteran's Administration (VA)
hospital in Northhampton Massachusetts which were used in a court case
United States versus Kristen Gilbert by Stephen Gehlbach, University of
Massachusetts, School of Public Health and Health Sciences. The data report
the number of deaths per year and shift in the hospital.
}
\keyword{datasets}
