% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/height_weight_data.R
\docType{data}
\name{height_weight}
\alias{height_weight}
\title{Height and weight data in a cross section}
\format{
A data frame with 846 rows and 6 variables:
\describe{
\item{age_at_death}{Country or Region, char}
\item{age}{Iso-code, char}
\item{state}{Year of measurement, num}
\item{sex}{1 male, 0 female, fct}
\item{height}{Height, num}
\item{weight}{Weight, num}
}
}
\source{
\url{https://github.com/rmcelreath/rethinking},
}
\usage{
height_weight
}
\description{
Data on from textbook on on Bayesian data analysis:
McElreath 2020. Statistical Rethinking, 2nd edition, CRC Press.
}
\keyword{datasets}
