% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean-height-women.R
\docType{data}
\name{height_women}
\alias{height_women}
\title{Mean height of women aged 18 or older by birth year, extending from 1896 to 1996.}
\format{
A data frame with 21008 rows and 6 variables:
\describe{
\item{Entity}{Country or Region, char}
\item{Code}{Iso-code, char}
\item{Year}{Year of measurement, num}
\item{Height}{height, num}
}
}
\source{
\url{https://ourworldindata.org/human-height}, \url{http://www.ncdrisc.org/data-downloads-height.html}
}
\usage{
height_women
}
\description{
Data on average height of women on different countries and regions in the world from 1896 - 1996
}
\keyword{datasets}
