% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life_expectancy.R
\docType{data}
\name{life_expectancy}
\alias{life_expectancy}
\title{Life expectancy and GDP per capita in 2011 international dollars.}
\format{
A data frame with 12447 rows and 5 variables:
\describe{
\item{Entity}{Country name, chr}
\item{Code}{ISO3 code of country, chr}
\item{Year}{Year, num, year of observation}
\item{LE}{Life expectancy at birth, num}
\item{GDPC}{GDP per capita in 2011 international dollars}
}
}
\source{
\url{https://ourworldindata.org/life-expectancy}
}
\usage{
life_expectancy
}
\description{
Data on life expectancy and GDP per capita in 2011 international dollars
for countries around the globe from 1950 -2018
}
\keyword{datasets}
