% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geysir_erruption_times.R
\docType{data}
\name{oldfaithful}
\alias{oldfaithful}
\title{Data on eruption times of the Geysir old faithful in Yellowstone
National Park, Wyoming USA}
\format{
A data frame with 272 rows and 1 variables:
\describe{
\item{time}{erruption time in minutes, num}

}
}
\source{
https://github.com/awstringer1/sta238-book/tree/master
}
\usage{
oldfaithful
}
\description{
This is a famous dataset on the erruption times of the Geysir
Old Faithful in Yellowstone National Park US. It is part of the R packages
of datasets where it is stored under the name faithful. We us a version from
the book of https://github.com/awstringer1/sta238-book/tree/master documenting
erruption times in minutes only
}
\keyword{datasets}
