% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pearson_father_son.R
\docType{data}
\name{pearson}
\alias{pearson}
\title{Historical data of Pearson 1903 of heights data of fathers and sons.}
\format{
A data frame with 1078 rows and 2 variables:
\describe{
\item{Father}{Height of father in inches, num}
\item{Son}{Height of son in inches, num}
}
}
\source{
Department of Statistics, University of California, Berkeley
\url{https://www.randomservices.org/random/data/Pearson.html}
}
\usage{
pearson
}
\description{
Heights of fathers and their sons in inches, based on a famous experiment by Karl Pearson
around 1903. The number of cases is 1078.
Random noise was added to the original data, to produce heights to the nearest 0.1 inch.
}
\keyword{datasets}
