% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/socr_height_weight.R
\docType{data}
\name{socr_height_weight}
\alias{socr_height_weight}
\title{Height and weight of a sample of boys and girls from SOCR.}
\format{
A data frame with 25000 rows and 3 variables:
\describe{
\item{Index}{Index of individual, num}
\item{Height}{Height in inches, num}
\item{Weight}{Weight in pounds, num}
}
}
\source{
\url{http://wiki.stat.ucla.edu/socr/index.php/SOCR_Data_Dinov_020108_HeightsWeights}
}
\usage{
socr_height_weight
}
\description{
SOCR Data - 25,000 Records of Human Heights (in) and Weights (lbs)
Human Height and Weight
Human Height and Weight are mostly hereditable, but lifestyles, diet,
health and environmental factors also play a role in determining
individual's physical characteristics.
}
\details{
The dataset below contains 25,000 synthetic records
of human heights and weights of 18 years old children.
These data were simulated based on a 1993 by a
Growth Survey of 25,000 children from birth to 18 years
of age recruited from Maternal and Child Health Centres (MCHC)
and schools and were used to develop Hong Kong's current
growth charts for weight, height, weight-for-age, weight-for-height
and body mass index (BMI). See also the Major League Baseball Players Height and Weight dataset.
}
\keyword{datasets}
